+++ example/messaging/app.dart
@@ -10,8 +10,8 @@ String contentUrl = '';
 main() {
   // xtag is null until the end of the event loop (known dart web ui issue)
   new Timer(0, (t) {
-    var webview = document.query('x-webview').xtag;
-    webview.on['loadcommit'].add(onLoadCommit);
+    var webview = document.query('x-webview').xtag;    
+    Webview.onLoadCommit.forTarget(webview).listen(onLoadCommit);
         
     // TODO(rms): We would like to build a content url of the form:
     // 'chrome-extension://<extensionID>/<pathToFile>' as described: